unwind dependencies
