+++ packages/search-common/package.json
@@ -36,17 +36,10 @@
     "url": "https://github.com/backstage/backstage/issues"
   },
   "dependencies": {
-    "@backstage/config": "^0.1.3",
-    "@backstage/plugin-search-indexer-backend": "^0.1.1"
+    "@backstage/config": "^0.1.3"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.6.2",
-    "@types/fs-extra": "^9.0.5",
-    "@types/git-url-parse": "^9.0.0",
-    "@types/js-yaml": "^3.12.5",
-    "@types/mime-types": "^2.1.0",
-    "@types/mock-fs": "^4.13.0",
-    "@types/recursive-readdir": "^2.2.0"
+    "@backstage/cli": "^0.6.2"
   },
   "jest": {
     "roots": [
