+++ plugins/search-backend/package.json
@@ -20,21 +20,17 @@
   },
   "dependencies": {
     "@backstage/backend-common": "^0.5.3",
-    "@backstage/config": "^0.1.2",
-    "@backstage/plugin-search-indexer-backend": "^0.1.1",
     "@backstage/search-common": "^0.1.1",
     "@types/express": "^4.17.6",
     "express": "^4.17.1",
     "express-promise-router": "^3.0.3",
     "winston": "^3.2.1",
-    "cross-fetch": "^3.0.6",
     "yn": "^4.0.0"
   },
   "devDependencies": {
     "@backstage/cli": "^0.6.0",
     "@types/supertest": "^2.0.8",
-    "supertest": "^4.0.2",
-    "msw": "^0.21.2"
+    "supertest": "^4.0.2"
   },
   "files": [
     "dist"
