+++ plugins/search-indexer-backend/src/registry.ts
@@ -17,6 +17,8 @@
 import {
   IndexableDocumentCollator,
   IndexableDocumentDecorator,
+} from '@backstage/search-common';
+import {
   RegisterCollatorParameters,
   RegisterDecoratorParameters,
 } from './types';
