+++ plugins/search-indexer-backend/src/types.ts
@@ -14,51 +14,10 @@
  * limitations under the License.
  */
 
-/**
- * Base properties that all indexed documents must include, as well as some
- * common properties that documents are encouraged to use where appropriate.
- */
-export interface IndexableDocument {
-  /**
-   * The primary name of the document (e.g. name, title, identifier, etc).
-   */
-  title: string;
-
-  /**
-   * Free-form text of the document (e.g. description, content, etc).
-   */
-  text: string;
-
-  /**
-   * The relative or absolute URL of the document (target when a search result
-   * is clicked).
-   */
-  location: string;
-
-  /**
-   * (Optional) The owner of the document (e.g. spec.owner on a catalog entity).
-   */
-  owner?: string;
-
-  /**
-   * (Optional) The lifecycle of the document (e.g. spec.lifecycle on a catalog entity).
-   */
-  lifecycle?: string;
-}
-
-/**
- * Signature for the callback function that implementors must register to have
- * their documents indexed.
- */
-export type IndexableDocumentCollator = () => Promise<IndexableDocument[]>;
-
-/**
- * Signature for the callback function that implementors must register to
- * decorate existing documents with additional metadata.
- */
-export type IndexableDocumentDecorator = (
-  documents: IndexableDocument[],
-) => Promise<IndexableDocument[]>;
+import {
+  IndexableDocumentCollator,
+  IndexableDocumentDecorator,
+} from '@backstage/search-common';
 
 /**
  * Parameters required to register a collator.
